requirements.txt
# Core
pydantic>=2.6
pyyaml>=6.0
python-dotenv>=1.0

# MQTT
paho-mqtt>=2.1.0

# HTTP (Frigate snapshot + health endpoint if needed)
httpx>=0.27

# OpenAI
openai>=1.30

# Utilities
tenacity>=8.2        # retry/backoff (snapshot + API calls)
Pillow>=10.0         # image resize/compress/crop preprocessing
pytest>=8.0          # test runner for integration and local workflows
